+++ a4kSubtitles/services/__init__.py
@@ -13,16 +13,7 @@ services = {}
 for service_name in __all__:
     service = services[service_name] = importlib.import_module('a4kSubtitles.services.%s' % service_name)
 
-    context = lambda: None
-
-    class Context:
-        def __getattr__(self, name):
-            return getattr(context, name, None)
-
-        def __setattr__(self, name, value):
-            return setattr(context, name, value)
-
-    service.context = Context()
+    service.context = utils.DictAsObject({})
 
     __set_fn_if_missing(service, 'build_auth_request', lambda _, __: None)